requirements.txt
# VRMonkey - Privacy Compliance Analysis for visionOS Apps
# Python 3.10+ required

# Data science & analysis
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.0.0

# Graph analysis
networkx>=2.8.0

# Computer vision
opencv-python>=4.7.0
Pillow>=9.4.0
mss>=9.0.0

# Web & API
Flask>=2.2.0
requests>=2.28.0
gradio-client>=0.7.0

# UI automation
pyautogui>=0.9.53
pynput>=1.7.6

# Hardware communication
pyserial>=3.5

# Data formats
PyYAML>=6.0

# macOS-specific (required for VRMonkey on macOS)
# pyobjc-framework-Cocoa>=9.0
# pyobjc-framework-Quartz>=9.0
